refact: #236 refact updateFriendPostList into useMutateFriendPostList
