+++ frontend/src/components/_common/question-item/QuestionItem.jsx
@@ -19,7 +19,7 @@ import {
   PostItemButtonsWrapper
 } from '@styles/wrappers';
 import { useTranslation } from 'react-i18next';
-import { updatePostsOnDeletePost } from 'src/queries/posts/updateFriendPostList';
+import useMutateFriendPostList from 'src/queries/posts/useMutateFriendPostList';
 import QuestionSendModal from './question-send-modal/QuestionSendModal';
 import { QuestionItemWrapper, Question } from './QuestionItem.styles';
 import NewResponse from './new-response/NewResponse';
@@ -71,13 +71,13 @@ export default function QuestionItem({ questionObj, onSubmit }) {
   const handleModalClose = () => {
     setQuestionSendModalOpen(false);
   };
-
+  const { mutateOnDeletePost } = useMutateFriendPostList();
   const handleDelete = () => {
     dispatch(deletePost(questionObj.id, questionObj.type));
 
     // FIXME: react-query 확장 적용할 때마다 업데이트 필요
     if (location.pathname === '/home') {
-      updatePostsOnDeletePost(questionObj.type, questionObj.id);
+      mutateOnDeletePost(questionObj.type, questionObj.id);
     }
   };
